--- a/d2l-course-tile.html
+++ b/d2l-course-tile.html
@@ -65,37 +65,39 @@ the user has in that organization - student, teacher, TA, etc.
 				</div>
 			</a>
 
-			<div class="hover-menu no-tap-interaction">
-				<d2l-dropdown>
-					<button	class="menu-item no-tap-interaction d2l-dropdown-opener" aria-label$="[[_courseSettingsLabel]]">
-						<d2l-icon icon="d2l-tier1:more"></d2l-icon>
-					</button>
-					<d2l-dropdown-menu id="overflow-dropdown">
-						<d2l-menu label$="[[_courseSettingsLabel]]">
-							<template is="dom-if" if="{{_courseInfoUrl}}" restamp="true">
-								<d2l-menu-item-link id="see-course-info-link"
-									text="{{localize('courseOfferingInformation')}}"
-									href="{{_courseInfoUrl}}">
-								</d2l-menu-item-link>
-							</template>
-							<template is="dom-if" if="{{_canChangeCourseImage}}" restamp="true">
-								<d2l-menu-item id="change-image-button"
-									text="{{localize('changeImage')}}"
-									on-d2l-menu-item-select="_launchCourseTileImageSelector">
+			<template is="dom-if" if="[[_showHoverMenu]]">
+				<div class="hover-menu no-tap-interaction">
+					<d2l-dropdown>
+						<button	class="menu-item no-tap-interaction d2l-dropdown-opener" aria-label$="[[_courseSettingsLabel]]">
+							<d2l-icon icon="d2l-tier1:more"></d2l-icon>
+						</button>
+						<d2l-dropdown-menu id="overflow-dropdown">
+							<d2l-menu label$="[[_courseSettingsLabel]]">
+								<template is="dom-if" if="{{_courseInfoUrl}}" restamp="true">
+									<d2l-menu-item-link id="see-course-info-link"
+										text="{{localize('courseOfferingInformation')}}"
+										href="{{_courseInfoUrl}}">
+									</d2l-menu-item-link>
+								</template>
+								<template is="dom-if" if="{{_canChangeCourseImage}}" restamp="true">
+									<d2l-menu-item id="change-image-button"
+										text="{{localize('changeImage')}}"
+										on-d2l-menu-item-select="_launchCourseTileImageSelector">
+									</d2l-menu-item>
+								</template>
+								<d2l-menu-item id="pin-button"
+									text$="[[_pinLabel]]"
+									on-d2l-menu-item-select="_hoverPinClickHandler"
+									on-focus="_hoverPinMenuItem"
+									on-blur="_unhoverPinMenuItem"
+									on-mouseover="_hoverPinMenuItem"
+									on-mouseout="_unhoverPinMenuItem">
 								</d2l-menu-item>
-							</template>
-							<d2l-menu-item id="pin-button"
-								text$="[[_pinLabel]]"
-								on-d2l-menu-item-select="_hoverPinClickHandler"
-								on-focus="_hoverPinMenuItem"
-								on-blur="_unhoverPinMenuItem"
-								on-mouseover="_hoverPinMenuItem"
-								on-mouseout="_unhoverPinMenuItem">
-							</d2l-menu-item>
-						</d2l-menu>
-					</d2l-dropdown-menu>
-				</d2l-dropdown>
-			</div>
+							</d2l-menu>
+						</d2l-dropdown-menu>
+					</d2l-dropdown>
+				</div>
+			</template>
 		</div>
 	</template>
 
@@ -180,6 +182,10 @@ the user has in that organization - student, teacher, TA, etc.
 				_notificationTitle: String,
 				_notificationDate: String,
 				isStartedInactive: {
+					type: Boolean,
+					value: false
+				},
+				_showHoverMenu: {
 					type: Boolean,
 					value: false
 				}
@@ -393,6 +399,7 @@ the user has in that organization - student, teacher, TA, etc.
 				});
 			},
 			_hoverCourseTile: function() {
+				this._showHoverMenu = true;
 				if (this.hoverEnabled) {
 					this._setCourseTileHovered(true);
 				}