Spiritual successor to US80997.

There are still a few places where changes can be made to improve the overall performance and intelligence of our usage of WCs within the widget. In particular, the d2l-course-tile WC is a candidate for some changes, as it is the most commonly-used by far within the widget. Initial testing with polyperf showed an average of 2000ms (range 1800-2250ms) to create 250 d2l-course-tile components, which is very high. In comparison, doing the same test with d2l-ajax takes around 200ms, an order of magnitude less. Obviously the course tile is much more complex than d2l-ajax, but the goal of these performance improvements is to potentially defer/avoid some of that complexity, as it will have an effect on page load times.

See US for details, but four things will be done in this PR:

- Remove creation of `d2l-ajax` elements until they are needed
- Use `afterNextRender`
- Don't create the `d2l-dropdown-menu` until the course tile is hovered
- Don't re-create d2l-siren-parser elements, instead create it once and reuse it

## Acceptance criteria:

- Unfortunately, outside of polyperf, it's difficult to quantify the performance improvements of web components within a page (and even the results within polyperf can be inconsistent). It is possible that some improvement may be seen in D2L.Performance.timing on a page that loads the My Courses widget, though - ultimately we would want page.display to be reduced some.
- No change to UI/UX should be noticeable

## Polyperf Results:

- Element created 250 times per run
- 25 runs
- using one standard deviation for average value
- average reduced:
  - IE: ~7100ms ➡️ ~3800ms
  - Chrome: ~1300ms ➡️ ~700ms
  - FF: ~3100ms ➡️ ~1300ms

### Before (Chrome)

![image](https://cloud.githubusercontent.com/assets/6231623/22316380/d6040bfa-e33a-11e6-8a3a-9624c0b1b4af.png)

### After (Chrome)

![image](https://cloud.githubusercontent.com/assets/6231623/22316327/6794f18e-e33a-11e6-8438-50f10797b33c.png)