@ryantmer 2017-01-20
Yay, finally got around to finishing this one off. Since this is entirely performance-focused (i.e. no behaviour should change), I didn't see a need for unit tests (aside from trivial things like "doesn't initially render with d2l-ajax elements", but that's of questionable value) - but it's entirely possible that some would be good, just in a way I didn't think of!