@ctwomey1 2017-01-24
I think this is a great first cut, but i think we can go deeper.  Based on how google seems to be using iron ajax i think we can cut the elements to just 1 for the entire application, instead of 2 for every ajax call per course tile

[Example](https://github.com/PolymerElements/app-localize-behavior/blob/master/app-localize-behavior.html#L183)

If i'm understanding this correctly they set up a static version of iron ajax on the prototype which is then shared between all of the mixins.  to utilize this properly we'd need to have a more programatic approach to building d2l-ajax allowing us to look at the promise and hook into it cleaner.