@ctwomey1 2017-01-24
Code LGTM, my suggestion doesn't block the current code from going in as it's still a large improvement