@trevorjonesd2l 2017-01-25
Test cases... verifying things still work.

- [ ] Pinned Courses load
- [ ] Overflow menus load
- [ ] Pin/Unpin action performs pinning/unpinning
- [ ] Change Image button opens change image overlay
- [ ] Course Offer link navigates to the course offering page
- [ ] Unpinned Courses page properly in all courses overlay
- [ ] Fills drop-down menu with auto-search results
- [ ] Live updates tile on search
- [ ] Filtering filters search results
- [ ] All Semesters load
- [ ] Can search for semesters
- [ ] All Departments load
- [ ] Can search for departments
- [ ] Sorting sorts the search results
- [ ] Telemetry requests are sent successfully