@ryantmer 2017-01-26
Updated the initial description with some polyperf details, showing what kind of effect this had on performance, so they can be reproduced if desired.