@ryantmer 2017-01-27
Just so @ctwomey1's earlier comment isn't lost, I've opened it as https://github.com/Brightspace/d2l-my-courses-ui/issues/285.